+++ example/workbook-host/src/client.rs
@@ -1,7 +1,6 @@
 use std::convert::Infallible;
 use postcard_rpc::{
-    host_client::{HostClient, HostErr},
-    standard_icd::{WireError, ERROR_PATH},
+    header::VarSeqKind, host_client::{HostClient, HostErr}, standard_icd::{WireError, ERROR_PATH}
 };
 use workbook_icd::{AccelRange, BadPositionError, GetUniqueIdEndpoint, PingEndpoint, Rgb8, SetAllLedEndpoint, SetSingleLedEndpoint, SingleLed, StartAccel, StartAccelerationEndpoint, StopAccelerationEndpoint};
 
@@ -39,8 +38,12 @@ impl<T, E> FlattenErr for Result<T, E> {
 
 impl WorkbookClient {
     pub fn new() -> Self {
-        let client =
-            HostClient::new_raw_nusb(|d| d.product_string() == Some("ov-twin"), ERROR_PATH, 8);
+        let client = HostClient::new_raw_nusb(
+            |d| d.product_string() == Some("ov-twin"),
+            ERROR_PATH,
+            8,
+            VarSeqKind::Seq2,
+        );
         Self { client }
     }
